+++ lending/src/sdk/models/shared/billcreditnotelineitem.ts
@@ -6,6 +6,7 @@ import { SpeakeasyBase, SpeakeasyMetadata } from "../../../internal/utils";
 import { AccountRef } from "./accountref";
 import { AccountsPayableTracking } from "./accountspayabletracking";
 import { TrackingCategoryRef } from "./trackingcategoryref";
+import { Zero } from "./zero";
 import { Expose, Type } from "class-transformer";
 
 /**
@@ -72,6 +73,18 @@ export class BillCreditNoteLineItem extends SpeakeasyBase {
     @Expose({ name: "accountRef" })
     @Type(() => AccountRef)
     accountRef?: AccountRef;
+
+    /**
+     * Links the current record line to the underlying record line that created it.
+     *
+     * @remarks
+     *
+     * For example, if a bill is generated from a purchase order, this property allows you to connect the bill line item to the purchase order line item in our data model.
+     */
+    @SpeakeasyMetadata()
+    @Expose({ name: "createdFromLineRef" })
+    @Type(() => Zero)
+    createdFromLineRef?: Zero;
 
     /**
      * Friendly name of each line item. For example, the goods or service for which credit has been received.