+++ lending/src/sdk/models/shared/index.ts
@@ -282,3 +282,4 @@ export * from "./validation";
 export * from "./validationitem";
 export * from "./weblink";
 export * from "./withholdingtaxitems";
+export * from "./zero";